Init Commit FloatAndDouble

+++ java-masterclass/ByteShortLongInt_Challenge/src/com/company/Main.java
@@ -5,11 +5,11 @@ public class Main {
     public static void main(String[] args) {
 //Challenge
         //Create a byte variable and set it to any valid byte number.
-        byte myByte = (byte) (100);
+        byte myByte = 100;
         System.out.println("myByte = " + myByte);
 
         //Create a short variable and set it to any valid short number.
-        short myShort = (short) (24000);
+        short myShort = 24000;
         System.out.println("myShort = " + myShort);
 
         //Create an int variable and set it to any valid int number.
@@ -17,7 +17,7 @@ public class Main {
         System.out.println("myInt = " + myInt);
 
         //Create a variable type long and make it equal to 50000 + 10 * (myByte + myShort + myInt)
-        long myLong = (long) (50000 + 10 * (myByte + myShort + myInt));
+        long myLong = (long) (50000L + 10L * (myByte + myShort + myInt));
         System.out.println("myLong = " + myLong);
     }
 }
